refactor(core): organize directory
