--- packages/core/.gitignore
@@ -1,2 +1,0 @@
-/components/
-/style/
