+++ packages/core/components/App.vue
