+++ packages/core/components/page/DefaultLayout.vue
