+++ packages/core/components/page/EndPage.vue
