+++ packages/core/package.json
@@ -6,7 +6,9 @@
   "main": "dist/index.mjs",
   "types": "dist/index.d.ts",
   "files": [
-    "dist/"
+    "dist/",
+    "components/",
+    "style/"
   ],
   "scripts": {
     "build": "unbuild"
