+++ packages/core/src/config.ts
