+++ packages/core/src/index.ts
@@ -1,1 +1,3 @@
-export * from './utils'
+export * from './config'
+export * from './page'
+export * from './theme'
