--- packages/core/src/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './config'
-export * from './pages'
-export * from './theme'
